fix(mobile): condition in the dockerfile to skip building

+++ apps/mobileExpo/build.sh
@@ -1,9 +1,10 @@
 #!/bin/sh
 
-npm install
-
-rm -rf dist/*
-
-npm install --global eas-cli
-
-eas build --local --non-interactive --platform android --profile production --output dist/area.apk
+if [ "$1" = "0" ]; then
+    npm install
+    rm -rf dist/*
+    npm install --global eas-cli
+    eas build --local --non-interactive --platform android --profile production --output dist/area.apk
+else
+    echo "Skipping mobile compilation..."
+fi
